# docs/swagger.yaml
basePath: /api
definitions:
...
  models.ProductProduct:
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@b-resto.com
    name: Soporte API
  description: API RESTful para sistema POS de restaurantes
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: B-Resto API
  version: "1.0"
paths:
...
  /product-variants:
    get:
      consumes:
      - application/json
      description: Obtiene lista de todas las variantes/SKUs de productos
      parameters:
      - description: Filtrar por template
        in: query
        name: template_id
        type: integer
      - description: Filtrar por estado activo
        enum:
        - "true"
        - "false"
        in: query
        name: is_active
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'data: array de product variants'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: mensaje'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Listar variantes de producto
      tags:
      - product-variants
    post:
      consumes:
      - application/json
      description: Crea una nueva variante/SKU de producto
      parameters:
      - description: Datos de la variante
        in: body
        name: variant
        required: true
        schema:
          $ref: '#/definitions/models.ProductProduct'
      produces:
      - application/json
      responses:
        "201":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Crear variante de producto
      tags:
      - product-variants
  /product-variants/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una variante/SKU (soft delete)
      parameters:
      - description: ID de la variante
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: Product variant deleted successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Product variant not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Eliminar variante de producto
      tags:
      - product-variants
    get:
      consumes:
      - application/json
      description: Obtiene una variante/SKU por ID
      parameters:
      - description: ID de la variante
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'data: product variant'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Product variant not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obtener variante de producto
      tags:
      - product-variants
    put:
      consumes:
      - application/json
      description: Actualiza los datos de una variante/SKU existente
      parameters:
      - description: ID de la variante
        in: path
        name: id
        required: true
        type: integer
      - description: Datos actualizados
        in: body
        name: variant
        required: true
        schema:
          $ref: '#/definitions/models.ProductProduct'
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: validación'
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 'error: Product variant not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Actualizar variante de producto
      tags:
      - product-variants
  /product-variants/{id}/toggle:
    patch:
      consumes:
      - application/json
      description: Cambia el estado is_active de una variante de producto
      parameters:
      - description: ID de la variante
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: message y data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: Product variant not found'
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Activar/Desactivar variante
      tags:
      - product-variants
...
securityDefinitions:
  Bearer:
    description: Escribir "Bearer {token}" para autenticación JWT
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"